refactor(desktop): move session preset domain to entities
