+++ apps/desktop/src/entities/session-preset/model/types.ts
@@ -1,0 +1,8 @@
+export type {
+  SaveSessionPresetPayload,
+  SaveTimeStepPresetPayload,
+  SessionPreset,
+  SessionPresetStepDraft,
+  SessionStepPreset,
+  TimeStepPreset,
+} from '@/shared/types';
