+++ apps/desktop/src/entities/session-preset/ui/index.ts
@@ -1,0 +1,4 @@
+export { AutoTagPicker } from './AutoTagPicker';
+export type { AutoTagPickerProps } from './AutoTagPicker';
+export { SessionPresetStepEditor } from './SessionPresetStepEditor';
+export type { SessionPresetStepEditorProps } from './SessionPresetStepEditor';
