+++ apps/desktop/src/entities/session-preset/ui/session-preset-step-editor.css
