+++ apps/desktop/src/features/croquis/ui/AutoTagPicker.tsx
@@ -1,192 +1,2 @@
-import { useEffect, useMemo, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Button, Chip, ChipButton, Modal, ModalFooter } from '../../../shared/ui';
-import type { Tag, TagGroup } from '../../../shared/types';
-import { TagSearchSelect } from '@/entities/tag';
-import './session-preset-step-editor.css';
-
-type AutoTagPickerProps = {
-  label: string;
-  tags: readonly Tag[];
-  availableTags: readonly Tag[];
-  tagGroups?: readonly TagGroup[];
-  disabled?: boolean;
-  emptyLabel?: string;
-  onTagAdd?: (tag: Tag) => void;
-  onTagRemove?: (tagId: string) => void;
-};
-
-export function AutoTagPicker({
-  label,
-  tags,
-  availableTags,
-  tagGroups = [],
-  disabled = false,
-  emptyLabel,
-  onTagAdd,
-  onTagRemove,
-}: AutoTagPickerProps) {
-  const { t } = useTranslation('common');
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [selectedTagId, setSelectedTagId] = useState('');
-  const canEditTags = Boolean(onTagAdd && onTagRemove);
-  const linkedTagIds = useMemo(() => new Set(tags.map(tag => tag.id)), [tags]);
-  const selectableTags = useMemo(
-    () => availableTags.filter(tag => !linkedTagIds.has(tag.id)),
-    [availableTags, linkedTagIds],
-  );
-  const selectedTag = useMemo(
-    () => selectableTags.find(tag => tag.id === selectedTagId) ?? null,
-    [selectableTags, selectedTagId],
-  );
-  const resolvedEmptyLabel =
-    emptyLabel ?? t('croquis.auto_tags.empty', { defaultValue: 'No auto tags' });
-  const tagEmptyMessage =
-    availableTags.length > 0 && selectableTags.length === 0
-      ? t('croquis.auto_tags.all_linked', { defaultValue: 'All tags are linked' })
-      : t('tags.no_tags_found', { defaultValue: 'No tags found' });
-
-  useEffect(() => {
-    if (selectedTagId && !selectableTags.some(tag => tag.id === selectedTagId)) {
-      setSelectedTagId('');
-    }
-  }, [selectableTags, selectedTagId]);
-
-  useEffect(() => {
-    if (!isModalOpen) {
-      return;
-    }
-
-    const handleEscape = (event: globalThis.KeyboardEvent) => {
-      if (event.key !== 'Escape') {
-        return;
-      }
-
-      event.preventDefault();
-      event.stopImmediatePropagation();
-      setIsModalOpen(false);
-      setSelectedTagId('');
-    };
-
-    window.addEventListener('keydown', handleEscape, { capture: true });
-    return () => {
-      window.removeEventListener('keydown', handleEscape, { capture: true });
-    };
-  }, [isModalOpen]);
-
-  const handleModalClose = () => {
-    setIsModalOpen(false);
-    setSelectedTagId('');
-  };
-
-  const handleAddTag = () => {
-    if (disabled || !selectedTag || !onTagAdd) {
-      return;
-    }
-
-    onTagAdd(selectedTag);
-    handleModalClose();
-  };
-
-  return (
-    <section className="session-preset-step-editor__group">
-      <span className="session-preset-step-editor__label">{label}</span>
-      <div className="session-preset-step-editor__tag-row">
-        {tags.map(tag =>
-          canEditTags ? (
-            <ChipButton
-              key={tag.id}
-              shape="rounded"
-              variant="neutral-dismiss"
-              disabled={disabled}
-              aria-label={t('tags.remove_tag', {
-                tag: tag.name,
-                defaultValue: 'Remove tag {{tag}}',
-              })}
-              onClick={() => {
-                onTagRemove?.(tag.id);
-              }}
-            >
-              {tag.name}
-            </ChipButton>
-          ) : (
-            <Chip key={tag.id} shape="rounded" variant="accent-outline">
-              {tag.name}
-            </Chip>
-          ),
-        )}
-        {tags.length === 0 ? (
-          <Chip shape="rounded" variant="accent-outline">
-            {resolvedEmptyLabel}
-          </Chip>
-        ) : null}
-        {canEditTags ? (
-          <ChipButton
-            shape="rounded"
-            variant="add"
-            disabled={disabled || selectableTags.length === 0}
-            aria-label={t('common.add_label', {
-              label,
-              defaultValue: 'Add {{label}}',
-            })}
-            onClick={() => {
-              setIsModalOpen(true);
-            }}
-          >
-            {t('tags.tag', { defaultValue: 'Tag' })}
-          </ChipButton>
-        ) : null}
-      </div>
-      {canEditTags && availableTags.length === 0 ? (
-        <span className="session-preset-step-editor__tag-hint">
-          {t('croquis.auto_tags.create_tags_first', {
-            defaultValue: 'Create tags in Tag Settings first.',
-          })}
-        </span>
-      ) : null}
-      <Modal
-        open={isModalOpen}
-        size="sm"
-        title={t('common.add_label', {
-          label,
-          defaultValue: 'Add {{label}}',
-        })}
-        onClose={handleModalClose}
-        closeOnEscape={false}
-        dialogClassName="session-preset-step-editor__tag-modal"
-        bodyClassName="session-preset-step-editor__tag-modal-body"
-        footer={
-          <ModalFooter alignment="end">
-            <Button size="sm" variant="secondary" onClick={handleModalClose}>
-              {t('common.cancel', { defaultValue: 'Cancel' })}
-            </Button>
-            <Button size="sm" disabled={disabled || selectedTag === null} onClick={handleAddTag}>
-              {t('common.add_tag', { defaultValue: 'Add Tag' })}
-            </Button>
-          </ModalFooter>
-        }
-      >
-        <TagSearchSelect
-          tags={selectableTags}
-          groups={tagGroups}
-          value={selectedTagId}
-          placeholder={t('tags.search_existing', { defaultValue: 'Search existing tags' })}
-          emptyMessage={tagEmptyMessage}
-          disabled={disabled || availableTags.length === 0}
-          aria-label={t('common.search_label', {
-            label,
-            defaultValue: 'Search {{label}}',
-          })}
-          onValueChange={nextTagId => {
-            setSelectedTagId(nextTagId);
-          }}
-        />
-        <span className="session-preset-step-editor__tag-hint">
-          {t('croquis.auto_tags.existing_only_hint', {
-            defaultValue: 'Existing tags only. Use Tag Settings to create or edit tags.',
-          })}
-        </span>
-      </Modal>
-    </section>
-  );
-}
+export { AutoTagPicker } from '@/entities/session-preset';
+export type { AutoTagPickerProps } from '@/entities/session-preset';
